+++ shardingsphere-features/shardingsphere-db-discovery/shardingsphere-db-discovery-core/src/main/java/org/apache/shardingsphere/dbdiscovery/distsql/exception/DatabaseDiscoveryRuleNotExistedException.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.dbdiscovery.distsql.exception;
 
-import org.apache.shardingsphere.infra.exception.RuleDefinitionViolationException;
+import org.apache.shardingsphere.infra.exception.rule.RuleDefinitionViolationException;
 
 import java.util.Collection;
