+++ shardingsphere-features/shardingsphere-readwrite-splitting/shardingsphere-readwrite-splitting-core/src/main/java/org/apache/shardingsphere/readwritesplitting/distsql/exception/InvalidLoadBalancersException.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.readwritesplitting.distsql.exception;
 
-import org.apache.shardingsphere.infra.exception.RuleDefinitionViolationException;
+import org.apache.shardingsphere.infra.exception.rule.RuleDefinitionViolationException;
 
 import java.util.Collection;
