+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-core/src/main/java/org/apache/shardingsphere/sharding/distsql/exception/ShardingBindingTableRuleNotExistsException.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.sharding.distsql.exception;
 
-import org.apache.shardingsphere.infra.exception.RuleDefinitionViolationException;
+import org.apache.shardingsphere.infra.exception.rule.RuleDefinitionViolationException;
 
 /**
  * Sharding binding table rule not exists exception.
